Remove highly specific patterns from gitignore
This is part of a redesign of the general gitignore structure.
Before, the gitignore was highly cluttered with specialized
patterns for specific things/environments because I tried
to make it handle all cases.
But now, I am making the gitignore as minimalistic as possible and
leaving it to individual repositories and environments to determine
which specialized patterns are needed.