Part 1 of editing readme
This is done to add context to additional branches.
Experiment more with what happens when you have a line between each
bullet entry vs when you don't.

Also make sure to split this commit because there is also a small edit
done to fit the rename of the setup guide.